requirements.txt
fastapi
uvicorn
python-multipart
python-pptx
pypdf
requests
